Added migration that add the "google_analytics_code" setting in the database.

+++ src/application/migrations/001_specific_calendar_sync.php
@@ -2,11 +2,11 @@
 
 /* ----------------------------------------------------------------------------
  * Easy!Appointments - Open Source Web Scheduler
- * 
+ *
  * @package     EasyAppointments
  * @author      A.Tselegidis <alextselegidis@gmail.com>
  * @copyright   Copyright (c) 2013 - 2015, Alex Tselegidis
- * @license     http://opensource.org/licenses/GPL-3.0 - GPLv3 
+ * @license     http://opensource.org/licenses/GPL-3.0 - GPLv3
  * @link        http://easyappointments.org
  * @since       v1.0.0
  * ---------------------------------------------------------------------------- */
@@ -14,17 +14,21 @@
 class Migration_Specific_calendar_sync extends CI_Migration {
 
 	public function up() {
-		$fields = array(
-			'google_calendar' => array(
-				'type' => 'VARCHAR',
-				'constraint' => '128',
-				'null' => TRUE)
-		);
-		
-		$this->dbforge->add_column('ea_user_settings', $fields);
+        if (!$this->db->field_exists('google_calendar', 'ea_user_settings')) {
+            $fields = array(
+                'google_calendar' => array(
+                    'type' => 'VARCHAR',
+                    'constraint' => '128',
+                    'null' => TRUE
+                )
+            );
+            $this->dbforge->add_column('ea_user_settings', $fields);
+        }
 	}
 
 	public function down() {
-		$this->dbforge->drop_column('ea_user_settings', 'google_calendar');
+        if ($this->db->field_exists('google_calendar', 'ea_user_settings')) {
+		  $this->dbforge->drop_column('ea_user_settings', 'google_calendar');
+        }
 	}
 }
